requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.0
pip==23.2.1
pydantic==2.0.3
pydantic_core==2.3.0
pydantic-settings==2.0.1
Pygments==2.15.1
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
requests==2.31.0
rich==13.4.2
setuptools==68.0.0
sniffio==1.3.0
SQLAlchemy==1.4.39
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
wheel==0.40.0
zipp==3.16.2

httpx~=0.24.1